# common/src/main/resources/api/openapi-spec.yaml
...
openapi: 3.0.3
info:
  title: Stocks
  version: 1.0.0
  license:
    name: GNU GPL v3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  description: |
    # Manage your household's stocks

    #### Versions

    Each entity in the system has a version which is incremented each time the
    entity is edited. This means all modifying operations have to pass the correct
    version of the entity to edit. If the version does not match an error is
    returned using status codes, see "Status Codes" for a list of status codes
    to support.

    #### Valid Time

    Each entity in the database is tagged with information about when it was present
    in the given state. This interval of validity is encoded using two timestamps,
    `validTimeStart` and `validTimeEnd`. The interval contains the starting
    timestamp while the ending timestamp is the first moment which is no longer
    included (closed-open notation, `[start, end)` using mathematical notation).

    #### Transaction Time

    Each entity in the database is tagged with information about when the
    information of the entity was entered into the database. This allows for
    reconstruction of past states of the database. This interval of validity is
    encoded using two timestamps, `transactionTimeStart` and `transactionTimeEnd`.
    The interval contains the starting timestamp while the ending timestamp is the
    first moment which is no longer included (closed-open notation, `[start, end)`
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
security:
  - type:
      - mutualTLS
...
components:
  schemas:
    id:
      type: integer
      minimum: 1
      description: the primary identifier of the entity
    version:
      type: integer
      minimum: 0
      description: the current version of the entity
    Response:
      type: object
      properties:
        status:
          type: integer
          minimum: 0
          maximum: 9
          description: See general description, section "Status Codes"
    Data:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
        version:
          $ref: '#/components/schemas/version'
    BitemporalData:
      allOf:
        - $ref: '#/components/schemas/Data'
        - type: object
          properties:
            initiates:
              type: integer
              minimum: 1
              description: |
                A reference to the UserDevice at the point in time when this
                entity was considered current.
            validTimeStart:
              type: string
              description: |
                The start of the time (inclusively) during which this entity is
                valid. Timestamp in the usual format.
            validTimeEnd:
              type: string
              description: |
                The end of the time (exclusively) during which this entity is
                valid. Timestamp in the usual format.
            transactionTimeStart:
              type: string
              description: |
                The start of the time (inclusively) during which this entity is
                part of the most recent truth on the data.
            transactionTimeEnd:
              type: string
              description: |
                The end of the time (exclusively) during which this entity is
                part of the most recent truth on the data.
...
    Unit:
      type: object
      properties:
        name:
          type: string
          description: The name of the unit
        abbreviation:
          type: string
          description: The abbreviation of the unit
...
  parameters:
    id:
      name: id
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/id'
      description: the primary identifier of the entity
    version:
      name: version
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/version'
      description: the version of the entity on which the request operates
    startingFrom:
      name: startingFrom
      in: query
      required: true
      schema:
        type: string
      description: |
        A timestamp stating that a GET response shall only include entities
        younger or equal to this value in transaction time. This is used to only
        get the latest changes to a table relative to an earlier query.
    upUntil:
      name: upUntil
      in: query
      required: true
      schema:
        type: string
      description: |
        A timestamp stating that a GET response shall only include entities
        older or equal to this value in transaction time. This is used to only
        get the latest changes to a table relative to an earlier query.
...
  responses:
    "200-write":
      description: Data modification successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
...
    "GET-400":
      description: No `startingFrom` or `upUntil` given.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
...
    "PUT-400":
      description: |
        * Invalid name given (absent or empty)
        * The entity was requested for update in the wrong version
        * Invalid input (see parameter limits)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
    "DELETE-400":
      description: |
        * This entity cannot be deleted because some other entity refers to it,
          e.g. a user still having devices or a location still containing food.
        * The entity was requested for deletion in the wrong version
        * Invalid input (see parameter limits)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
...
    "404":
      description: |
        * No entity with this id found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
    "500":
      description: |
        * CA unreachable (User device operations only)
        * Database unreachable
        * General error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
paths:
...
  /v2/unit:
    put:
      parameters:
        - name: name
          in: query
          schema:
            type: string
            minLength: 1
        - name: abbreviation
          in: query
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          $ref: '#/components/responses/200-write'
        "400":
          $ref: '#/components/responses/PUT-400'
        "500":
          $ref: '#/components/responses/500'
    get:
      parameters:
        - $ref: '#/components/parameters/startingFrom'
        - $ref: '#/components/parameters/upUntil'
      responses:
        "200":
          description: Return all units of measure of the system
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          oneOf:
                            - allOf:
                                - $ref: '#/components/schemas/Unit'
                                - $ref: '#/components/schemas/BitemporalData'
                            - allOf:
                                - $ref: '#/components/schemas/Unit'
                                - $ref: '#/components/schemas/Data'
        "400":
          $ref: '#/components/responses/GET-400'
        "500":
          $ref: '#/components/responses/500'
    delete:
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/version'
      responses:
        "200":
          $ref: '#/components/responses/200-write'
        "400":
          $ref: '#/components/responses/DELETE-400'
        "404":
          $ref: '#/components/responses/404'
        "500":
          $ref: '#/components/responses/500'
  /v2/unit/rename:
    put:
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/version'
        - name: name
          in: query
          schema:
            type: string
            minLength: 1
        - name: abbreviation
          in: query
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          $ref: '#/components/responses/200-write'
        "400":
          $ref: '#/components/responses/PUT-400'
        "404":
          $ref: '#/components/responses/404'
        "500":
          $ref: '#/components/responses/500'
...